SQL tests cleanup
