+++ hazelcast-sql/src/test/java/com/hazelcast/jet/sql/impl/OptimizerInstantiationTest.java
@@ -16,39 +16,26 @@
 
 package com.hazelcast.jet.sql.impl;
 
-import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.instance.impl.HazelcastInstanceProxy;
-import com.hazelcast.sql.impl.optimizer.SqlOptimizer;
+import com.hazelcast.jet.sql.SqlTestSupport;
 import com.hazelcast.sql.impl.SqlServiceImpl;
-import com.hazelcast.sql.impl.SqlTestSupport;
-import com.hazelcast.test.HazelcastParallelClassRunner;
-import com.hazelcast.test.TestHazelcastInstanceFactory;
-import com.hazelcast.test.annotation.ParallelJVMTest;
-import com.hazelcast.test.annotation.QuickTest;
-import org.junit.AfterClass;
+import com.hazelcast.sql.impl.optimizer.SqlOptimizer;
+import org.junit.BeforeClass;
 import org.junit.Test;
-import org.junit.experimental.categories.Category;
-import org.junit.runner.RunWith;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
-@RunWith(HazelcastParallelClassRunner.class)
-@Category({QuickTest.class, ParallelJVMTest.class})
 public class OptimizerInstantiationTest extends SqlTestSupport {
 
-    private static final TestHazelcastInstanceFactory FACTORY = new TestHazelcastInstanceFactory(1);
-
-    @AfterClass
-    public static void afterClass() {
-        FACTORY.shutdownAll();
+    @BeforeClass
+    public static void setUpClass() {
+        initialize(1, null);
     }
 
     @Test
     public void testCreate() {
-        HazelcastInstance instance = FACTORY.newHazelcastInstance();
-
-        SqlServiceImpl service = ((HazelcastInstanceProxy) instance).getOriginal().node.getNodeEngine().getSqlService();
+        SqlServiceImpl service = ((HazelcastInstanceProxy) instance()).getOriginal().node.getNodeEngine().getSqlService();
 
         SqlOptimizer optimizer = service.getOptimizer();
