+++ hazelcast-sql/src/test/java/com/hazelcast/jet/sql/impl/index/IndexFilterTestSupport.java
@@ -16,8 +16,8 @@
 
 package com.hazelcast.jet.sql.impl.index;
 
+import com.hazelcast.jet.sql.SqlTestSupport;
 import com.hazelcast.query.impl.CompositeValue;
-import com.hazelcast.sql.impl.SqlTestSupport;
 import com.hazelcast.sql.impl.exec.scan.index.IndexFilterValue;
 import com.hazelcast.sql.impl.expression.ConstantExpression;
 import com.hazelcast.sql.impl.type.QueryDataType;
@@ -29,7 +29,8 @@ import java.util.Collections;
  * Utility stuff for index filter tests.
  */
 @SuppressWarnings("rawtypes")
-public class IndexFilterTestSupport extends SqlTestSupport {
+public abstract class IndexFilterTestSupport extends SqlTestSupport {
+
     public static ConstantExpression constant(Object value, QueryDataType type) {
         return ConstantExpression.create(value, type);
     }
