+++ hazelcast-sql/src/test/java/com/hazelcast/jet/sql/impl/optimizer/PlanKeyTest.java
@@ -16,20 +16,14 @@
 
 package com.hazelcast.jet.sql.impl.optimizer;
 
-import com.hazelcast.sql.impl.SqlTestSupport;
+import com.hazelcast.jet.sql.SqlTestSupport;
 import com.hazelcast.sql.impl.optimizer.PlanKey;
-import com.hazelcast.test.HazelcastParallelClassRunner;
-import com.hazelcast.test.annotation.ParallelJVMTest;
-import com.hazelcast.test.annotation.QuickTest;
 import org.junit.Test;
-import org.junit.experimental.categories.Category;
-import org.junit.runner.RunWith;
 
 import static java.util.Collections.singletonList;
 
-@RunWith(HazelcastParallelClassRunner.class)
-@Category({QuickTest.class, ParallelJVMTest.class})
 public class PlanKeyTest extends SqlTestSupport {
+
     @Test
     public void testEquals() {
         PlanKey key = new PlanKey(singletonList(singletonList("schema1")), "sql1");
